requirements.txt
supabase
python-dotenv
pandas
numpy
sqlalchemy